nlp_core/requirements.txt
spacy
transformers
pdfminer.six
python-docx
jsonschema
ifcopenshell
rapidfuzz
PyPDF2